deploy-orchestrator/requirements.txt
# Azure Functions Core Requirements (Minimal stable versions)
azure-functions==1.18.0
requests==2.31.0

# PDF Processing (Basic versions)
PyPDF2==3.0.1
reportlab==4.0.4

# OpenAI Integration (Stable version)
openai>=1.52.0

# Essential Azure Services only
azure-storage-blob>=12.14.0
azure-identity>=1.13.0
azure-cosmos>=4.5.0

# Document Intelligence (if needed)
azure-ai-formrecognizer>=3.3.0
azure-cognitiveservices-vision-computervision>=0.9.0

# Removed azure-ai-projects (Bing Search Grounding not working)

# Basic multi-agent dependencies (minimal versions)
dataclasses-json>=0.6.0
pydantic>=2.5.0

# Web scraping essentials
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Core dependencies
python-dotenv>=1.0.0

# Last updated: Mon Aug 25 23:29:05 CEST 2025
